Rewards Deletion Enhanced

+++ src/main/java/com/rewards/backend/app/referral/referral/RewardController.java
@@ -1,14 +1,23 @@
 package com.rewards.backend.app.referral.referral;
 
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import com.rewards.backend.ResponseHandler;
 import com.rewards.backend.exception.CustomException;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/rewards")
@@ -76,15 +85,19 @@ public class RewardController {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Object> deleteRewardPlan(@PathVariable("id") long id) {
-    	try {
-            rewardPlanRepository.deleteById(id);
-            return ResponseHandler.generateResponse("Successfully deleted", HttpStatus.OK, "OK");
-		}
-    	catch (Exception e) {
-			return ResponseHandler.generateResponse(e.getMessage(), HttpStatus.BAD_REQUEST, "Failed to load data");
-		}
-    	
+        try {
+            Optional<RewardPlan> rewardPlanOptional = rewardPlanRepository.findById(id);
+            if (rewardPlanOptional.isPresent()) {
+                rewardPlanRepository.deleteById(id);
+                return ResponseHandler.generateResponse("Successfully deleted", HttpStatus.OK, "OK");
+            } else {
+                return ResponseHandler.generateResponse("Content not found", HttpStatus.NOT_FOUND, "Content not found");
+            }
+        } catch (Exception e) {
+            return ResponseHandler.generateResponse(e.getMessage(), HttpStatus.BAD_REQUEST, "Failed to delete reward plan");
+        }
     }
+
 
     @PostMapping("/addMultiple")
     public ResponseEntity<Object> addMultipleRewards(@RequestBody List<RewardPlan> rewardPlans) {
